+++ model/modules/__init__.py
@@ -1,15 +1,19 @@
-from .transformer import TransformerBlock, CausalSelfAttention, FeedForward
-from .vision_encoder import VisionEncoder, PatchEmbedding
-from .diffusion import DiffusionModule, UNet, NoiseScheduler
-from .video_generator import VideoGenerator, TemporalAttention
 from .audio_encoder import AudioEncoder, AudioFrontEnd
 from .audio_generator import AudioGenerator
+from .diffusion import DiffusionModule, NoiseScheduler, UNet
+from .moe import MoEFeedForward, build_ffn
 from .multimodal_projector import MultimodalProjector
+from .transformer import CausalSelfAttention, FeedForward, RMSNorm, TransformerBlock
+from .video_generator import TemporalAttention, VideoGenerator
+from .vision_encoder import PatchEmbedding, VisionEncoder
 
 __all__ = [
     "TransformerBlock",
     "CausalSelfAttention",
     "FeedForward",
+    "RMSNorm",
+    "MoEFeedForward",
+    "build_ffn",
     "VisionEncoder",
     "PatchEmbedding",
     "DiffusionModule",